Package into an application, plus tweaks
