+++ .gitignore
@@ -1,1 +1,5 @@
 *.pyc
+*.egg-info
+cache/
+build/
+dist/
