+++ setup.cfg
@@ -1,0 +1,56 @@
+# -*- coding: utf-8 -*-
+[metadata]
+name = pidi-spotify
+version = 0.0.1
+author = Philip Howard
+author_email = phil@pimoroni.com
+description = Utility for displaying Spotify track data and album art alongside Raspotify/Librespot
+long_description = file: README.md
+long_description_content_type = text/markdown
+keywords = Raspberry Pi
+url = https://www.pimoroni.com
+project_urls =
+	GitHub=https://www.github.com/pimoroni/pidi-spotify
+license = MIT
+# This includes the license file(s) in the wheel.
+# https://wheel.readthedocs.io/en/stable/user_guide.html#including-license-files-in-the-generated-wheel-file
+license_files = LICENSE.txt
+classifiers =
+	Development Status :: 4 - Beta
+	Operating System :: POSIX :: Linux
+	License :: OSI Approved :: MIT License
+	Intended Audience :: Developers
+	Programming Language :: Python :: 3
+	Programming Language :: Python :: 3.7
+	Programming Language :: Python :: 3.8
+	Topic :: Multimedia
+	Topic :: System :: Hardware
+
+[options]
+include_package_data = True
+python_requires = >= 3.7
+packages = pidi_spotify
+install_requires =
+	pidi_display_st7789
+	spotipy
+
+[options.entry_points]
+console_scripts =
+	pidi_spotify = pidi_spotify:main
+
+[flake8]
+exclude =
+	.tox,
+	.eggs,
+	.git,
+	__pycache__,
+	build,
+	dist
+ignore =
+	E501
+
+[pimoroni]
+py2deps =
+py3deps =
+configtxt =
+commands =
