+++ start.sh
@@ -5,4 +5,4 @@ cd /home/pi/pidi-spotify
 export SPOTIPY_CLIENT_ID="YOUR CLIENT ID"
 export SPOTIPY_CLIENT_SECRET="YOUR CLIENT SECRET"
 
-python3 -m pidi_spotify
+pidi_spotify
